% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_width.R
\name{bottom_width}
\alias{bottom_width}
\title{Compute Bottom Width}
\usage{
bottom_width(bfw, wd, z)
}
\arguments{
\item{bfw}{Vector of bankfull width or active channel widths.}

\item{wd}{Vector of width to depth ratio (W/D) values.}

\item{z}{Vector of channel angle Z values.}
}
\value{
vector of bottom widths
}
\description{
Same procedure as used in Heat Source 7. If the width:depth ratio and
channel angle z values produces bottom width < 0 z is incrementally decreased
using the formula: z_new = z * 0.99 until bottom width >= 0.
}
