% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.7dadm.R
\name{calc.7dadm}
\alias{calc.7dadm}
\title{Calculate the 7DADM from hourly data.}
\usage{
calc.7dadm(
  df,
  datetime_col = "Datetime",
  sim_col = "sim",
  stream_km_col = "Stream_km",
  value_col = "value"
)
}
\arguments{
\item{df}{The input dataframe. df must have the following columns names:
Datetime (with datetime as POSIXlt),
sim (with simulation name in character format),
Stream_km (as numeric),
value (numeric hourly stream temperature)}

\item{datetime_col}{Datetime column name. Default is 'Datetime'.}

\item{sim_col}{Simulation column name. Default is 'sim'.}

\item{stream_km_col}{Stream km column name. Default is 'Stream_km'.}

\item{value_col}{Value column name. Default is 'value'.}
}
\value{
dataframe
}
\description{
Calculate the daily maximum and rolling seven day average of the daily maximums (7DADM) by date from hourly data.
The rolling average result is right aligned and indexed to the last day of the seven day period.
Constituent and statistic names are added as an ID variable.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
