% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hs7.porosity.R
\name{calc.hs7.porosity}
\alias{calc.hs7.porosity}
\title{Calculate porosity}
\usage{
calc.hs7.porosity(bed_particle_size, embeddedness, fine_particle_size = 0.062)
}
\arguments{
\item{bed_particle_size}{Bed particle size, in millimeters.}

\item{embeddedness}{Embeddedness, as a decimal fraction from 0 to 1.}

\item{fine_particle_size}{Fine sediment particle size, in millimeters. Default
is 0.062 mm, based on value used in Heat Source 7 code.}
}
\value{
Porosity, as a decimal fraction from 0 to 1.
}
\description{
Calculates porosity using heat source 7 channel morphology parameters.
The porosity equation is adopted from Bedient and Huber 1992, as described
in the heat source 7 user manual. This function is commonly used to convert
heat source 7 input parameters to heat source 8 or 9 porosity inputs.
}
