% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hs7.sed.conductivity.R
\name{calc.hs7.sed.conductivity}
\alias{calc.hs7.sed.conductivity}
\title{Calculate sediment thermal conductivity}
\usage{
calc.hs7.sed.conductivity(
  bed_particle_size_mm,
  embeddedness,
  fine_particle_size_mm = 0.062,
  sediment_density = 1600,
  water_density = 1000,
  sediment_heat_capacity = 2219,
  water_heat_capacity = 4187,
  sediment_thermal_diffusivity_m2_s = 4.5e-06,
  water_thermal_diffusivity_m2_s = 1.4331e-07
)
}
\arguments{
\item{bed_particle_size_mm}{Bed particle size, in millimeters.}

\item{embeddedness}{Embeddedness, as a decimal fraction from 0 to 1.}

\item{fine_particle_size_mm}{Fine sediment particle size, in millimeters.
Default is 0.062 mm, based on value used in Heat Source 7 code.}

\item{sediment_density}{Sediment density, in kilograms per cubic meter.
Default is 1600 kg/m^3, based on value used in Heat Source 7 code.}

\item{water_density}{Water density, in kilograms per cubic meter.
Default is 1000 kg/m^3, based on value used in Heat Source 7 code.}

\item{sediment_heat_capacity}{Sediment specific heat capacity, in joules
per kilogram per degree Celsius. Default is 2219 J/(kg deg C),
based on value used in Heat Source 7 code.}

\item{water_heat_capacity}{Water specific heat capacity, in joules per
kilogram per degree Celsius. Default is 4187 J/(kg deg C), based on value
used in Heat Source 7 code.}

\item{sediment_thermal_diffusivity_m2_s}{Sediment thermal diffusivity,
in square meters per second. Default is 0.0000045 m^2/s, based on value
used in Heat Source 7 code.}

\item{water_thermal_diffusivity_m2_s}{Water thermal diffusivity,
in square meters per second. Default is 0.00000014331 m^2/s, based on value
used in Heat Source 7 code.}
}
\value{
Sediment thermal conductivity, in watts per meter per degree Celsius.
}
\description{
Calculates sediment thermal conductivity using heat source 7 channel
morphology parameters. The calculation is the same as what is used in
heat source 7. This function is commonly used to convert
heat source 7 input parameters to heat source 8 or 9 sediment thermal
conductivity inputs.
}
