% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.hs7.sed.thickness.R
\name{calc.hs7.sed.thickness}
\alias{calc.hs7.sed.thickness}
\title{Calculate sediment hyporheic thickness}
\usage{
calc.hs7.sed.thickness(
  bed_particle_size,
  min_thickness = 0.1,
  max_thickness = 1
)
}
\arguments{
\item{bed_particle_size}{Bed particle size, in millimeters.}

\item{min_thickness}{Minimum sediment hyporheic thickness, in meters.
Default is 0.1 m, based on value used in Heat Source 7 code.}

\item{max_thickness}{Maximum sediment hyporheic thickness, in meters.
Default is 1 m, based on value used in Heat Source 7 code.}
}
\value{
Sediment hyporheic thickness, in meters.
}
\description{
Calculates sediment hyporheic thickness using heat source 7 channel
morphology parameters. The calculation is the same as what is used in
heat source 7. Heat source 7 bounds the thickness from 0.1 meters to 1 meter.
This function is commonly  used to convert heat source 7 input parameters to
heat source 8 or 9 sediment hyporheic thickness inputs.
}
