% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_7dadm.R
\name{calc_7dadm}
\alias{calc_7dadm}
\title{Calculate the 7DADM from hourly temperature data.}
\usage{
calc_7dadm(
  df,
  datetime_col = "datetime",
  sim_col = "sim",
  stream_km_col = "stream_km",
  value_col = "value"
)
}
\arguments{
\item{datetime_col}{datetime column name. Default is 'datetime'.}

\item{sim_col}{Simulation column name. Default is 'sim'.}

\item{stream_km_col}{Stream km column name. Default is 'stream_km'.}

\item{value_col}{Value column name. Default is 'value'.}
}
\value{
data frame with columns 'sim', 'datetime', 'date', 'stream_km', 'constituent', and 'value'.
}
\description{
Calculate the daily maximum and rolling seven day average of the daily maximums (7DADM) by date from hourly data.
The rolling average result is right aligned and indexed to the last day of the seven day period.
'Daily Maximum Temperature' and '7DADM Temperature' are values added to the
constituent column.
}
