% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gof.R
\name{calc_gof}
\alias{calc_gof}
\title{Calculate goodness of fit stats}
\usage{
calc_gof(preds, obs, tolerance = 60)
}
\arguments{
\item{preds}{Data frame of model prediction results. The \code{preds} data
      frame must include the following columns:
\itemize{
\item Monitoring.Location.ID: Character value of the \code{obs} monitoring
      location ID
\item constituent: Character value of the constituent (e.g. 'Temperature,
      water').
\item datetime: POSIXct datetime.
\item model_km: numeric model stream kilometer
\item value: Constituent value.
}}

\item{obs}{Data frame of field observation results. The \code{obs} data
      frame must include the following columns:
\itemize{
\item Monitoring.Location.ID: Character value of the \code{obs} monitoring
      location ID
\item constituent: Character value of the constituent (e.g. 'Temperature,
      water').
\item datetime: POSIXct datetime.
\item model_km: numeric model stream kilometer
\item value: Constituent value.
}}

\item{tolerance}{The maximum number of minutes allowed between the datetimes
in \code{preds} and \code{obs} to establish a match. If the difference
exceeds the tolerance the prediction is not include in the goodness
of fit stats. The tolerance cannot be greater than 1440 minutes
(1 day). The default tolerance is 30 minutes.}
}
\value{
data frame
}
\description{
Returns a data frame of calculated goodness of fit stats for each
monitoring site and all sites combined. The function accepts multiple sites
and variable timeseries. \code{preds} and \code{obs} are matched based on the
Monitoring.Location.ID, constituent, and nearest datetime within a set number
of minutes defined using \code{tolerance}.If the difference in datetimes
exceeds the \code{tolerance} a match will not be made. Goodness of fit
statistics are calculated using the \code{\link[hydroGOF]{hydroGOF}} package.
Statistics calculated are Mean Error (ME), Mean Absolute Error (MAE), Root
Mean Square Error (RMSE), Nash-Sutcliffe Efficiency (NSE), and  Coefficient
of Determination (R2). The count of \code{preds} and \code{obs} comparisons
at each monitoring location are also included (n).
}
\seealso{
\code{\link[hydroGOF]{me}}, \code{\link[hydroGOF]{mae}},
         \code{\link[hydroGOF]{rmse}}, \code{\link[hydroGOF]{NSE}},
         \code{\link[base]{cor}}
}
