% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.outputs.R
\name{format.outputs}
\alias{format.outputs}
\title{Format raw Heat Source output to a consistent format}
\usage{
\method{format}{outputs}(df, hs_ver = 9, sheet_name = NULL, constituent_name = NA, sim_name = NA)
}
\arguments{
\item{df}{The path to directory where the Heat Source 7 .xlsm model is located.}

\item{hs_ver}{The version of Heat Source. Input is a numeric value equal to 7, 8, or 9. Default is 9.}

\item{sheet_name}{The name of the Heat Source 7 output worksheet to read. String format. NULL for Heat Source 8-9. Default is NULL.}

\item{constituent_name}{The name of the output. String format. Optional. Default is NA. If NA, the function will attempt to look up the constituent using [lookup_output].}

\item{sim_name}{The name of the model scenario. String format. Optional. Default is NA.}
}
\value{
dataframe
}
\description{
Format output from any of the read.hs functions so the dataframe has a consistent structure and column names.
}
