% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_outputs.R
\name{format_outputs}
\alias{format_outputs}
\title{Format Heat Source output to a long format}
\usage{
format_outputs(df, hs_ver = 9, name = NA, constituent_name = NA, sim_name = NA)
}
\arguments{
\item{df}{A data frame output from any of the read.hs functions}

\item{hs_ver}{The version of Heat Source. Input is a numeric value
equal to 6, 7, 8, or 9. Default is 9.}

\item{name}{The name of the .txt, or .csv output file for Heat Source
version 8-9; or excel sheet name for Heat Source 6-7. String format.}

\item{constituent_name}{The name of the output. String format. Optional. Default is NA.
If NA, the function will attempt to look up the constituent using \code{\link{output_lookup}}
and the value in 'name'.}

\item{sim_name}{The name of the model scenario. String format. Optional. Default is NA.}
}
\value{
data frame
}
\description{
Format output from any of the read.hs functions so the returned data frame
is in long format with a consistent structure and column names.
}
\details{
Output column names include:
\itemize{
\item sim: Character value of the model scenario simulation name or other description.
\item constituent: Character value of the constituent (e.g. Temperature).
\item datetime: POSIXct datetime.
\item date: Character date in format "mm/dd/YYYY".
\item model_km: numeric stream kilometer extracted from the model.
\item value: Constituent value.
}

Heat Source 6 long distances are converted to stream kilometer with the most
downstream long distance = stream kilometer 0.
}
