% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcdata_buffer.R
\name{lcdata_buffer}
\alias{lcdata_buffer}
\title{Make a buffer width scenario.}
\usage{
lcdata_buffer(
  node_id = NA_real_,
  stream_id = NA_character_,
  aspect,
  skygap,
  vegwd_L,
  vegwd_R,
  lccode_L,
  lccode_R,
  lccode_noveg,
  lccode_stream,
  trans_dir = c(45, 90, 135, 180, 225, 270, 315, 360),
  transsample_count,
  transsample_distance
)
}
\arguments{
\item{node_id}{Unique heat source node ID. Optional. Default is NA.}

\item{stream_id}{Unique stream identifier. Optional. Default is NA.}

\item{aspect}{The stream aspect in degrees. Can be a vector.}

\item{skygap}{The distance (meters) between the left and right bank vegetation buffers where there is open sky.
This distance is about the same as the active channel width. Can take a vector of values as input.}

\item{vegwd_L}{The width of the vegetation buffer on the left bank (meters) looking downstream. Can take a vector of values as input.}

\item{vegwd_R}{The width of the vegetation buffer on the right bank (meters) looking downstream. Can take a vector of values as input.}

\item{lccode_L}{The numeric land cover code to use for samples that fall inside the vegetated buffer on the left bank. Do not use zero.}

\item{lccode_R}{The numeric land cover code to use for samples that fall inside the vegetated buffer on the right bank. Do not use zero.}

\item{lccode_noveg}{The numeric land cover code to use for samples that fall outside the vegetated buffer. Do not use zero.}

\item{lccode_stream}{The numeric land cover code to use for samples that fall on the stream. Do not use zero.}

\item{trans_dir}{A vector of integer values corresponding to the specific azimuth
directions of the sample transects. Default is Heat Source 9 directions = c(45, 90, 135, 180, 225, 270, 315, 360)}

\item{transsample_count}{Number of samples per transect. The number does not include the sample at the stream node.}

\item{transsample_distance}{The distance between transect samples (meters).}
}
\value{
dataframe
}
\description{
This function assists with preparing heat source land cover data input to evaluate
different buffer width scenarios.
}
\details{
The arguments \code{aspect}, \code{skygap}, \code{vegwd_L}, and \code{vegwd_R} can take a
vector of multiple values. When this is the case all combinations of the
supplied vectors are used to map the land cover codes. The returned data frame
will have a row for each unique combination.
}
