% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_near.R
\name{match_near}
\alias{match_near}
\title{Value matching based on closest value}
\usage{
match_near(x, y, tolerance = Inf, value = TRUE)
}
\arguments{
\item{x}{Numeric vector of the values to be matched.}

\item{y}{Numeric vector of the values to be matched against}

\item{tolerance}{The maximum numeric difference between values in \code{x}
and \code{y} to establish a match. If the difference exceeds the
tolerance NA is returned for that match. The default tolerance is Inf
which will return the closest value or position regardless of the
absolute difference.}

\item{value}{If TRUE, return the value of \code{y} instead of the position.
Default is TRUE.}
}
\value{
vector of values or positions from \code{y} that are closest to the
values in vector \code{x}.
}
\description{
\code{match_near} is similar to the base function
\code{\link[base]{match}} except \code{match_near} returns a vector of
values or positions from \code{y} that are closest to the the vector of
values in \code{x}. If two values in vector\code{y} are equally close to the
value in \code{x} the first one is used.
}
\details{
The closest match in \code{y} to a given value in \code{x} is determined
by the the value that has the minimum absolute difference between \code{x}
and \code{y}. e.g. \code{which.min(abs(x - y))}
}
