% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near_join.R
\name{near_join}
\alias{near_join}
\title{Left join based on the closest value}
\usage{
near_join(x, y, by = "datetime", tolerance)
}
\arguments{
\item{x}{Data frame of to index against}

\item{y}{Data frame that will be joined to \code{x} based on closest
value in x.}

\item{by}{The column name to join by. There can only be one column.
If x and y have different column names,
use the format \code{by = c("a" = "b")}. Default is "datetime".}

\item{tolerance}{The maximum numeric difference between values in \code{x}
and \code{y} to establish a match. If the difference exceeds the
tolerance NA is returned for that row in x. The default tolerance is Inf
which will return the closest value or position regardless of the
absolute difference.}
}
\description{
Add rows from y to x based on the closest numeric value in a single common
column. This function is used to join based on matches of values or datetimes
within a numeric tolerance.
}
