% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_lookup.R
\name{output_lookup}
\alias{output_lookup}
\title{Return a vector of the con.}
\usage{
output_lookup(file_name = NULL, hs_ver = 9, sheet_name = NULL)
}
\arguments{
\item{file_name}{The name of the .txt or .csv output file for Heat Source version 8-9.}

\item{hs_ver}{The version of Heat Source. Input is a numeric value equal to 7, 8, or 9. Default is 9.}

\item{sheet_name}{The name of the Heat Source 7 output worksheet. String format. NULL for Heat Source 8-9
the name of the .txt or .csv output file for Heat Source version 8-9.
For Heat Source 7, the ".xlsm" extension should be included in the file name.}
}
\value{
vector of values with the first element being the output constituent and the second the units.
}
\description{
This function will return the name of the output constituent and units based on the input file name or sheet.
}
