% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs.outputs.R
\name{read.hs.outputs}
\alias{read.hs.outputs}
\title{Read hourly Heat Source 7-8 outputs.}
\usage{
read.hs.outputs(
  output_dir,
  file_name,
  hs_ver = 9,
  sheet_name = NULL,
  constituent_name = NA,
  sim_name = NA
)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 7 .xlsm model is located.}

\item{file_name}{The file name of the .xlsm Heat Source 7 model or
the name of the .txt or .csv output file for Heat Source version 8-9.
For Heat Source 7, the ".xlsm" extension should be included in the file name.}

\item{hs_ver}{The version of Heat Source. Input is a numeric value equal to 7, 8, or 9. Default is 9.}

\item{sheet_name}{The name of the Heat Source 7 output worksheet to read. String format. NULL for Heat Source 8-9. Default is NULL.}

\item{constituent_name}{The name of the output. String format. Default is NA. If NA, the function will attempt to look up the constituent using [lookup_output].}

\item{sim_name}{The name of the model scenario. String format. Default is NA.}
}
\value{
data frame
}
\description{
Read hourly output from Heat Source version 7-9.
Heat Source 7 excel workbook needs to be saved as .xlsm. Workbooks in .xls do not work.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
