% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs6.flow.R
\name{read.hs6.flow}
\alias{read.hs6.flow}
\title{Read Heat Source 6 flow.}
\usage{
read.hs6.flow(output_dir, file_name, sheet_name = "Main Menu")
}
\arguments{
\item{output_dir}{The path to the directory where the Heat Source 6 .xlsm
model is located.}

\item{file_name}{The file name of the .xlsm Heat Source 6 model.
The ".xlsm" extension should be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read.}
}
\value{
data frame
}
\description{
Read the flow rates from a Heat Source version 6 model.
Heat Source excel workbook needs to be saved as .xlsm. Workbooks in .xls do
not seem to work.
}
\details{
Data is returned in wide format with the long distance used as the column
name for all values. An "X" is added as a prefix to every long distance value
to have syntactically valid column names. The datetime is the first column
and is formatted in excel numeric date format.

This function is setup to read the "Main Menu" worksheet.
}
\keyword{6}
\keyword{Heat}
\keyword{Source}
\keyword{version}
