% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs6.temp.R
\name{read.hs6.temp}
\alias{read.hs6.temp}
\title{Read Heat Source 6 longitudinal temperature outputs.}
\usage{
read.hs6.temp(output_dir, file_name, sheet_name = "Long Temp Output")
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 6 .xlsm model is located.}

\item{file_name}{The file name of the .xlsm Heat Source 6 model.
The ".xlsm" extension should be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read. Default is "Long Temp Output".}
}
\value{
data frame
}
\description{
Read hourly xlsm temperature output from Heat Source version 6.
Heat Source excel workbook needs to be saved as .xlsm. Workbooks in .xls do not work.
}
\details{
Data is returned in wide format with the long distance used as the column name for all values.
An "X" is added as a prefix to every long distance value to have syntactically valid
column names. The datetime is the first column and is formatted in excel
numeric date format.

This function is setup to read the "Long Temp Output" worksheet.
}
\keyword{6}
\keyword{Heat}
\keyword{Source}
\keyword{version}
