% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs7.outputs.R
\name{read.hs7.outputs}
\alias{read.hs7.outputs}
\title{Read Heat Source 7 outputs.}
\usage{
read.hs7.outputs(output_dir, file_name, sheet_name)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 7 .xlsm model is located.}

\item{file_name}{The file name of the .xlsm Heat Source 7 model.
The ".xlsm" extension should be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read.}
}
\value{
dataframe
}
\description{
Read hourly xlsm output from Heat Source version 7.
Heat Source excel workbook needs to be saved as .xlsm. Workbooks in .xls do not work.
}
\details{
Data is returned in wide format with the stream km used as the column name for all values.
An "X" is added as a prefix to every stream km value to have syntactically valid
column names. The datetime is the first column and is formatted in excel
numeric date format.

This function reads the following output sheets:
 \itemize{
 \item Output - Solar Potential
 \item Output - Solar Surface
 \item Output - Solar Received
 \item Output - Conduction
 \item Output - Longwave
 \item Output - Convection
 \item Output - Evaporation
 \item Output - Total Heat
 \item Output - Evaporation Rate
 \item Output - Temperature
}
To read heat source 7 effective shade outputs use \code{\link{read.hs7.shade}}.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
