% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs7.shade.R
\name{read.hs7.shade}
\alias{read.hs7.shade}
\title{Read Heat Source 7 effective shade outputs.}
\usage{
read.hs7.shade(
  output_dir,
  file_name,
  sim_name,
  constituent_name = "Effective Shade",
  statistic_name = "Percent",
  sheet_name = "Chart-Shade"
)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source .xlsx or output files are located.}

\item{file_name}{The file name of the .xlsx heat source model.
The ".xlsx" extension shoudl be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read.}

\item{...}{Other arguments passed on to readxl::read_excel().}
}
\value{
dataframe
}
\description{
Read xlsx effective shade output from Heat Source version 7.
Heat Source excel workbook needs to be saved as .xlsx. Workbooks in .xls do not seem to work.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
