% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs7.velocity.R
\name{read.hs7.velocity}
\alias{read.hs7.velocity}
\title{Read Heat Source 7 flow velocity outputs.}
\usage{
read.hs7.velocity(output_dir, file_name, sheet_name = "Output - Hydraulics")
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 7 .xlsm model
is located.}

\item{file_name}{The file name of the .xlsm Heat Source 7 model.
The ".xlsm" extension should be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read. Default is
'Output - Hydraulics'.}
}
\value{
dataframe
}
\description{
Read flow velocity outputs from the hydraulics output worksheet in a Heat
Source version 7 model. The Heat Source excel workbook needs to be saved as
.xlsm. Workbooks in .xls do not seem to work.
}
\details{
Data is returned in wide format with the model stream km used as the column
name for all values. An "X" is added as a prefix to every stream km value to
have syntactically valid column names. The datetime is the first column and
is formatted in excel numeric date format. Heat Source 7 has a bug where the
dates in the hydraulics output worksheet are one day off. This function works
around this bug and returns the correct date for each result.

This function is setup to read the "Output - Hydraulics" worksheet.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
