% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs8.outputs.R
\name{read.hs8.outputs}
\alias{read.hs8.outputs}
\title{Read Heat Source 8 outputs.}
\usage{
read.hs8.outputs(output_dir, file_name)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 8 output files are located.}

\item{file_name}{The file name of the Heat Source 8 output. The file is a .txt text file.}
}
\value{
data frame
}
\description{
Read hourly output from Heat Source version 8.
}
\keyword{8}
\keyword{Heat}
\keyword{Source}
\keyword{version}
