% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs9.outputs.R
\name{read.hs9.outputs}
\alias{read.hs9.outputs}
\title{Read Heat Source 9 outputs.}
\usage{
read.hs9.outputs(output_dir, file_name)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 9 output files are located.}

\item{file_name}{The file name of the Heat Source 9 output. Thee file is a csv text file.}
}
\value{
dataframe
}
\description{
Read output from Heat Source version 9.
}
\keyword{8}
\keyword{Heat}
\keyword{Source}
\keyword{version}
