% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.obs.R
\name{read.obs}
\alias{read.obs}
\title{Read field observation data}
\usage{
read.obs(obs_dir, file_name)
}
\arguments{
\item{obs_dir}{The path to directory where observation data .xlsx file is located.}

\item{file_name}{The name of the observation data .xlsx file.}
}
\description{
Read a .xlsx workbook formatted with field observation data. The workbook is 
setup using a similar format as DEQ's AWQMS data submission templates.
}
\details{
The workbook must contain two worksheets named:
\itemize{
 \item Monitoring_Locations
 \item Results
 }

The worksheets must contain the columns identified in the lists below.
Additional columns can be included but will not be used by the function.

Required 'Monitoring_Locations' worksheet columns
\itemize{
 \item Monitoring.Location.ID: Character value of the monitoring location ID.
 \item Latitude: Numeric Latitude of the monitoring location in decimal degrees.
 \item Longitude: Numeric Longitude of the monitoring location in decimal degrees.
 \item GNIS_Name: Character name of the waterbody.
 \item model_km: Numeric model stream kilometer corresponding to the location of the monitoring site
 }

Required 'Results' worksheet columns:
 \itemize{
 \item Monitoring.Location.ID: Character value of the monitoring location ID.
 \item Activity.Start.Date: Date of sample. Formatted as an excel date in the format 'YYYY/MM/DD'.
 \item Activity.Start.Time: Time of sample. Formatted as an excel date in the format 'HH24:MM'.
 \item Activity.Start.End.Time Zone: Character timezone (e.g. PDT or PST).
 \item Characteristic.Name: Characteristic name used in AWQMS. (e.g. 'Temperature, water', 'Flow') 
 \item Result.Value: Measured characteristic result value.
 \item Result.Unit: Units for measured characteristic result. Same as what is used in AWQMS. (e.g. 'deg C', 'm3/sec', 'm') 
 }
}
