% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_any.R
\name{round_any}
\alias{round_any}
\title{Round to multiple of any number}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{Numeric vector of the values to be rounded}

\item{accuracy}{number to round to}

\item{f}{Rounding function. One of: \code{\link{floor}}, \code{\link{ceiling}} or
\code{\link{round}}}
}
\description{
Same function from plyr.
}
\examples{
v <- c(1.22, 2.45, 4.12)
round_any(min(v, na.rm = TRUE), accuracy = 0.5 , f = floor)
round_any(max(v, na.rm = TRUE), accuracy = 0.5 , f = ceiling)
}
