% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatsourcetools-package.R
\docType{package}
\name{heatsourcetools-package}
\alias{heatsourcetools}
\alias{heatsourcetools-package}
\title{heatsourcetools: Heat Source Model Post Processing Tools}
\description{
R package to assist post processing of heat source models.
}
\author{
\strong{Maintainer}: Ryan Michie \email{Ryan.Michie@deq.oregon.gov}

}
\keyword{internal}
