% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2km.R
\name{obs2km}
\alias{obs2km}
\title{Matches observation location to the nearest model location}
\usage{
obs2km(obskm, simkm)
}
\arguments{
\item{obskm}{Numeric vector of the observation location (e.g. stream kilometer)}

\item{simkm}{Numeric vector of the model output locations (e.g. stream kilometer)}
}
\value{
vector equal to the length of obskm
}
\description{
Returns a vector of model kilometers (or distances) that are closest to the 
observation kilometer or distance. If an observation id is equally distant 
between two output kilometers the first one in the vector is used.
}
