% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs.outputs.R
\name{read.hs.outputs}
\alias{read.hs.outputs}
\title{Read hourly Heat Source 7-8 outputs.}
\usage{
read.hs.outputs(
  output_dir,
  file_name,
  hs_ver = 9,
  sheet_name = NULL,
  constituent_name = NULL,
  statistic_name = NULL,
  sim_name = NULL
)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 7 .xlsx model is located.}

\item{file_name}{The file name of the .xlsx Heat Source 7 model or
the name of the .txt or .csv output file for Heat Soruce version 8-9.
For Heat Soruce 7, the ".xlsx" extension should be included in the file name.}

\item{hs_ver}{The version of Heat Source. Input is a numeric value equal to 7, 8, or 9. Default is NULL.}

\item{sheet_name}{The name of the Heat Source 7 output worksheet to read. String format. NULL for Heat Soruce 8-9. Defualt is 9}

\item{constituent_name}{The name of the output. String format. Default is NULL}

\item{sim_name}{The name of the model scenerio. String format. Default is NULL}
}
\value{
dataframe
}
\description{
Read hourly output from Heat Source version 7-9.
Heat Source 7 excel workbook needs to be saved as .xlsx. Workbooks in .xls do not work.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
