% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs7.outputs.R
\name{read.hs7.outputs}
\alias{read.hs7.outputs}
\title{Read Heat Source 7 outputs.}
\usage{
read.hs7.outputs(output_dir, file_name, sheet_name)
}
\arguments{
\item{output_dir}{The path to directory where the Heat Source 7 .xlsx model is located.}

\item{file_name}{The file name of the .xlsx Heat Source 7 model.
The ".xlsx" extension should be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read.}

\item{...}{Other arguments passed on to readxl::read_excel().}
}
\value{
dataframe
}
\description{
Read hourly xlsx output from Heat Source version 7.
Heat Source excel workbook needs to be saved as .xlsx. Workbooks in .xls do not seem to work.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
