% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hs7.shade.R
\name{read.hs7.shade}
\alias{read.hs7.shade}
\title{Read Heat Source 7 effective shade outputs.}
\usage{
read.hs7.shade(output_dir, file_name, sheet_name = "Chart-Shade")
}
\arguments{
\item{output_dir}{The path to the directory where the Heat Source .xlsm is located.}

\item{file_name}{The file name of the .xlsm heat source model.
The ".xlsm" extension should be included in the file name.}

\item{sheet_name}{The name of the output worksheet to read. Default is "Chart-Shade".}

\item{...}{Other arguments passed on to readxl::read_excel().}
}
\value{
data frame
}
\description{
Read xlsm effective shade output from Heat Source version 7.
Heat Source excel workbook needs to be saved as .xlsm. Workbooks in .xls do not seem to work.
}
\keyword{7}
\keyword{Heat}
\keyword{Source}
\keyword{version}
